pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.13",
    "ffmpeg-python>=0.2.0",
    "google-api-python-client>=2.176.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "gtts>=2.5.4",
    "mutagen>=1.47.0",
    "openai>=1.93.3",
    "pillow>=11.3.0",
    "python-telegram-bot>=22.2",
]
